% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_df.R
\name{hash_df}
\alias{hash_df}
\title{Hash columns in a data frame}
\usage{
hash_df(df)
}
\arguments{
\item{df}{The data frame to be hashed.}
}
\value{
A dataframe with two columns: name and hash
}
\description{
The function hash_df accepts a data frame and performs an MD5 has on each column
using the openssl package. The original column names and md5 hash checksums are returned as a dataframe.
}
\details{
The function may be used in the process of determining which columns have changed in a data frame.
}
\examples{
hash_df(iris[1])
hash_df(iris)
}
