% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hash_df.R
\name{hash_df}
\alias{hash_df}
\title{Hash rows or columns in a data frame}
\usage{
hash_df(df, margin = c("rows", "columns"))
}
\arguments{
\item{df}{The data frame to be hashed.}

\item{margin}{A character value indicating the margin on which the hash will be calculated (rows or columns).}
}
\value{
A dataframe with two columns: name and hash
}
\description{
The function hash_df accepts a data frame and performs an MD5 along the specified margin using the openssl package.
When margin = "rows", the original data frame is return with an additional column labeled md5 containing row-level hash values.
When margin = "columns", the column names from the original data frame and md5 hash are returned as a data frame.
}
\details{
The function may be used in the process of determining which columns have changed in a data frame.
}
\examples{
hash_df(iris[1])
hash_df(iris)
}
